feat: create shapeHelper and move types in template.models
