+++ src/modules/templates/pages/builderPage/components/basics/shapes/shapeHelper.tsx
@@ -1,0 +1,16 @@
+import { CSSProperties } from 'react';
+import { SVGItemProps } from 'src/modules/templates/models/template.model';
+
+export const shapeHelper = {
+  convertColorsToCSSVariables: (colors: SVGItemProps['colors']) => {
+    if (!colors) {
+      return {};
+    }
+    const colorsCSSProperties = Object.fromEntries(
+      Object.entries(colors).map(([key, value]) => {
+        return [`--${key}`, value];
+      }),
+    ) as CSSProperties;
+    return colorsCSSProperties;
+  },
+};
